+++ src/pyubx2/socket_wrapper.py
@@ -1,5 +1,5 @@
 """
-socket_stream class.
+socket_wrapper class.
 
 A skeleton socket wrapper which provides basic stream-like
 read(bytes) and readline() methods.
@@ -19,9 +19,9 @@ Created on 4 Apr 2022
 from socket import socket
 
 
-class SocketStream:
+class SocketWrapper:
     """
-    socket stream class.
+    Socket wrapper class.
     """
 
     def __init__(self, sock: socket, **kwargs):