+++ src/pyubx2/ubxtypes_decodes.py
@@ -18,12 +18,14 @@ GNSSLIST = {
     6: "GLONASS",
     7: "NAVIC",
 }
+"""GNSS code"""
 
 # UBX-CFG-DGNSS
 DGNSMODE = {
     2: "RTK float",
     3: "RTK fixed",
 }
+"""Differential correction mode from UBX-CFG-DGNSS"""
 
 # UBX-CFG-GEOFENCE
 CONFLVL = {
@@ -33,6 +35,7 @@ CONFLVL = {
     3: "99.7%",
     4: "99.99%",
 }
+"""Confidence level from UBX-CFG-GEOFENCE"""
 
 # UBX-CFG-GNSS
 # key is (gnssId, sigCfMask)
@@ -55,12 +58,17 @@ SIGCFMASK = {
     (6, 0x01): "GLONASS L1",
     (6, 0x10): "GLONASS L2",
 }
+"""
+Signal & carrier frequency code from UBX-CFG_GNSS
+ - key is (gnssId, sigCfMask)
+"""
 
 # UBX-CFG-INF
 PROTOCOLID = {
     0: "UBX",
     1: "NMEA",
 }
+"""Protocol from UBX-CFG-INF"""
 
 # UBX-CFG-NAV5
 DYNMODEL = {
@@ -77,6 +85,7 @@ DYNMODEL = {
     11: "robotic lawn mower",  # (supported for protocol versions 33.21)
     12: "electric kick scooter",  # (supported for protocol versions 33.21, and 35.10)
 }
+"""Dynamic model from UBX-CFG-NAV5"""
 
 # UBX-CFG-NAV5
 FIXMODE = {
@@ -84,6 +93,7 @@ FIXMODE = {
     2: "3D only",
     3: "auto 2D/3D",
 }
+"""Fix mode from UBX-CFG-NAV5"""
 
 # UBX-CFG-NMEA
 NMEAVERSION = {
@@ -93,12 +103,14 @@ NMEAVERSION = {
     0x23: "NMEA version 2.3",
     0x21: "NMEA version 2.1",
 }
+"""NMEA version from UBX-CFG-NMEA"""
 
 # UBX-CFG-NMEA
 SVNUMBERING = {
     0: "Strict - Satellites are not output",
     1: "Extended - Use proprietary numbering",  # (see Satellite Numbering)
 }
+"""Space vehicle numbering from UBX-CFG-NMEA"""
 
 # UBX-CFG-NMEA
 MAINTALKERID = {
@@ -110,15 +122,18 @@ MAINTALKERID = {
     5: "Set main Talker ID to 'GB'",  # (not supported for protocol versions less than 15.00)
     6: "Set main Talker ID to 'GQ'",  # (available in NMEA 4.11 and later)
 }
+"""Main talker ID from UBX-CFG-NMEA"""
 
 # UBX-CFG-NMEA
 GSVTALKERID = {
     0: "Use GNSS-specific Talker ID (as defined byNMEA)",
     1: "Use the main Talker ID",
 }
+"""GSV talker ID from UBX-CFG-NMEA"""
 
 # UBX-CFG-ODO
 ODOPROFILE = {0: "running", 1: "cycling", 2: "swimming", 3: "car", 4: "custom"}
+"""Odometer profile from UBX-CFG-ODO"""
 
 # UBX-CFG-PRT
 CHARLEN = {
@@ -127,28 +142,37 @@ CHARLEN = {
     0x10: "7bit",  # (supported only with parity)
     0x11: "8bit",
 }
+"""Serial port character length from UBX-CFG-PRT"""
+
 PARITY = {
     "000": "Even parity",
     "001": "Odd parity",
     "10X": "No parity",
     "X1X": "Reserved",
 }
+"""Serial port parity from UBX-CFG-PRT"""
+
 NSTOPBITS = {
     0x00: "1 Stop bit",
     0x01: "1.5 Stop bit",
     0x10: "2 Stop bit",
     0x11: "0.5 Stop bit",
 }
+"""Serial port stop bits from IBX-CFG-PRT"""
+
 POL = {
     0: "High-active",
     1: "Low-active",
 }
+"""Serial port polarity from UBX-CFG-PRT"""
+
 SPIMODE = {
     0x00: "SPI Mode 0: CPOL = 0, CPHA = 0",
     0x01: "SPI Mode 1: CPOL = 0, CPHA = 1",
     0x10: "SPI Mode 2: CPOL = 1, CPHA = 0",
     0x11: "SPI Mode 3: CPOL = 1, CPHA = 1",
 }
+"""SPI mode from UBX-CFG-PRT"""
 
 # UBX-CFG-PWR
 STATE = {
@@ -156,6 +180,7 @@ STATE = {
     0x53544F50: "GNSS stopped",
     0x42434B50: "Software backup",  # USB interface will be disabled, other wakeup source is needed.
 }
+"""Power state from UBX-CFG-PWR"""
 
 # UBX-CFG-RATE
 TIMEREF = {
@@ -166,6 +191,7 @@ TIMEREF = {
     4: "Galileo time",  # (not supported for protocol versions less than 18.00)
     5: "NavIC time",  # (not supported for protocol versions less than 29.00)
 }
+"""Time reference from UBX-CFG-RATE"""
 
 # UBX-CFG-RST
 NAVBBRMASK = {
@@ -173,6 +199,8 @@ NAVBBRMASK = {
     0x0001: "Warm start",
     0xFFFF: "Cold start",
 }
+"""Navigation battery-backed RAM mask from UBX-CFG-RST"""
+
 RESETMODE = {
     0x00: "Hardware reset (watchdog) immediately",
     0x01: "Controlled software reset",
@@ -181,6 +209,7 @@ RESETMODE = {
     0x08: "Controlled GNSS stop",
     0x09: "Controlled GNSS start",
 }
+"""Reset mode from UBX-CFG-RST"""
 
 # UBX-CFG-TMODE3
 MODE = {
@@ -188,6 +217,7 @@ MODE = {
     1: "Survey In",
     2: "Fixed Mode",  # (true ARP position information required)
 }
+"""Timing mode from UBX-CFG-TMODE3"""
 
 # UBX-CFG-TP5
 GRIDUTCGNSS = {
@@ -197,6 +227,7 @@ GRIDUTCGNSS = {
     3: "BeiDou",
     4: "Galileo",  # (not supported for protocol versions less than 18.00)
 }
+"""Timing Grid UTC GNSS source from UBX-CFG-TP5"""
 
 # UBX-MON-COMMS
 PROTIDS = {
@@ -207,6 +238,7 @@ PROTIDS = {
     6: "SPARTN",
     0xFF: "No protocol reported",
 }
+"""Protocol IDs from UBX-MON-COMMS"""
 
 # UBX-MON-HW, UBX-MON-RF
 ASTATUS = {
@@ -216,17 +248,22 @@ ASTATUS = {
     3: "SHORT",
     4: "OPEN",
 }
+"""Antenna status from UBX-MON-HW, UBX-MON-RF"""
+
 APOWER = {
     0: "OFF",
     1: "ON",
     2: "DONTKNOW",
 }
+"""Antenna power from UBX-MON-HW, UBX-MON-RF"""
+
 JAMMINGSTATE = {
     0: "unknown or feature disabled",
     1: "ok - no significant jamming",
     2: "warning - interference visible but fix OK",
     3: "critical - interference visible and no fix",
 }
+"""Jamming state from UBX-MON-HW, UBX-MON-RF"""
 
 # UBX-MON-SYS
 BOOTTYPE = {
@@ -242,17 +279,21 @@ BOOTTYPE = {
     9: "VDD_RF fail",
     10: "V_CORE_HIGH fail",
 }
+"""Boot type from UBX-MON-SYS"""
 
 # UBX-NAV-GEOFENCE
 GEOFENCE_STATUS = {
     0: "Geofencing not available or not reliable",
     1: "Geofencing active",
 }
+"""Geofence status from UBX-NAV-GEOFENCE"""
+
 COMBSTATE = {
     0: "unknown",
     1: "inside",
     2: "outside",
 }
+"""Comb state from UBX-NAV-GEOFENCE"""
 
 # UBX-NAV-ORB
 HEALTH = {
@@ -260,12 +301,15 @@ HEALTH = {
     1: "healthy",
     2: "not healthy",
 }
+"""Space vehicle health from UBX-NAV-ORB"""
+
 VISIBILITY = {
     0: "unknown",
     1: "below horizon",
     2: "above horizon",
     3: "above elevation mask",
 }
+"""Space vehicle visibility from UBX-NAV-ORB"""
 
 # UBX-NAV-PL
 PLPOSFRAME = PLVELFRAME = {
@@ -274,6 +318,7 @@ PLPOSFRAME = PLVELFRAME = {
     2: "Longitudinal-Lateral-Vertical",
     3: "HorizSemiMajorAxis-HorizSemiMinorAxis-Vertical",
 }
+"""Position protection level from UBX-NAV-PL"""
 
 # UBX-NAV-PVT
 GPSFIX = FIXTYPE = {
@@ -284,6 +329,8 @@ GPSFIX = FIXTYPE = {
     4: "GPS + DR",
     5: "TIME ONLY",
 }
+"""Fix type from UBX-NAV-PVT"""
+
 PSMSTATE = {
     0: "PSM is not active",
     1: "Enabled",
@@ -292,11 +339,15 @@ PSMSTATE = {
     4: "Power Optimized Tracking",
     5: "Inactive",
 }
+"""Power save mode state from UBX-NAV-PVT"""
+
 CARRSOLN = {
     0: "NO RTK",
     1: "RTK FLOAT",
     2: "RTK FIXED",
 }
+"""Carrier phase range solution from UBX-NAV-PVT"""
+
 LASTCORRECTIONAGE = {
     0: 0,
     1: 1,
@@ -311,6 +362,7 @@ LASTCORRECTIONAGE = {
     10: 90,
     11: 120,
 }
+"""Last DGPS correction age from UBX-NAV-PVT"""
 
 # UBX-NAV-SAT
 QUALITYIND = {
@@ -323,6 +375,8 @@ QUALITYIND = {
     6: "code and carrier locked and time synchronized",
     7: "code and carrier locked and time synchronized",
 }
+"""Signal quality indicator from UBX-NAV-SAT"""
+
 ORBITSOURCE = {
     0: "no orbit information is available for this SV",
     1: "ephemeris is used",
@@ -333,6 +387,7 @@ ORBITSOURCE = {
     6: "other orbit information is used",
     7: "other orbit information is used",
 }
+"""Orbit source from UBX-NAV-SAT"""
 
 # UBX-NAV-SBAS
 SBASMODE = {
@@ -340,6 +395,8 @@ SBASMODE = {
     1: "Enabled integrity",
     3: "Enabled test mode",
 }
+"""SBAS (satellite-based augmentation system) mode from UBX-NAV-SBAS"""
+
 SBASSYS = {
     -1: "Unknown",
     0: "WAAS",
@@ -348,11 +405,14 @@ SBASSYS = {
     3: "GAGAN",
     16: "GPS",
 }
+"""SBAS system from UBX-NAV-SBAS"""
+
 SBASINTEGRITYUSED = {
     0: "Unknown",
     1: "Integrity information is not available or SBAS integrity is not enabled",
     2: "Receiver uses only GPS satellites for which integrity information is available",
 }
+"""SBAS integrity used indicator from UBX-NAV-SBAS"""
 
 # UBX-NAV-SIG
 CORRSOURCE = {
@@ -366,6 +426,8 @@ CORRSOURCE = {
     7: "SPARTN",
     8: "CLAS",
 }
+"""DGPS correction source from UBX-NAV-SIG"""
+
 IONOMODEL = {
     0: "none",
     1: "Klobuchar GPS",
@@ -373,6 +435,8 @@ IONOMODEL = {
     3: "Klobuchar BeiDou",
     4: "Dual Frequency",
 }
+"""Ionospheric model type from UBX-NAV-SIG"""
+
 # key is (gnssId, sigId)
 SIGID = {
     (0, 0): "GPS L1C/A",
@@ -403,29 +467,10 @@ SIGID = {
     (6, 2): "GLONASS L2 OF",
     (7, 0): "NavIC L5 A",
 }
-
-SOURCEOFCURLS = {
-    0: "Default",
-    1: "Derived from time difference between GPS and GLONASS time",
-    2: "GPS",
-    3: "SBAS",
-    4: "BeiDou",
-    5: "Galileo",
-    6: "Aided data",
-    7: "Configured",
-    8: "NavIC",
-    255: "Unknown",
-}
-
-SRCOFLSCHANGE = {
-    0: "No source",
-    2: "GPS",
-    3: "SBAS",
-    4: "BeiDou",
-    5: "Galileo",
-    6: "GLONASS",
-    7: "NavIC",
-}
+"""
+GNSS, Signal ID from UBX-NAV-SIG
+ - key is (gnssId, sigId)
+"""
 
 # UBX-NAV-STATUS
 SPOOFDETSTATE = {
@@ -434,12 +479,15 @@ SPOOFDETSTATE = {
     2: "spoofing indicated",
     3: "multiple spoofing indications",
 }
+"""Spoof detection state from UBX-NAV-STATUS"""
+
 PSMSTATUS = {
     0: "aquisition",
     1: "tracking",
     2: "power optimised tracking",
     3: "inactive",
 }
+"""Power save mode state from UBX-NAV-STATUS"""
 
 # UBX-NAV-TIME*
 UTCSTANDARD = {
@@ -454,3 +502,29 @@ UTCSTANDARD = {
     8: "National Physics Laboratory India (NPLI)",
     15: "Unknown",
 }
+"""UTC time standard from UBX-NAV-TIME*"""
+
+SOURCEOFCURLS = {
+    0: "Default",
+    1: "Derived from time difference between GPS and GLONASS time",
+    2: "GPS",
+    3: "SBAS",
+    4: "BeiDou",
+    5: "Galileo",
+    6: "Aided data",
+    7: "Configured",
+    8: "NavIC",
+    255: "Unknown",
+}
+"""Information source for the current number of leap seconds, from UBX-NAV-TIMELS"""
+
+SRCOFLSCHANGE = {
+    0: "No source",
+    2: "GPS",
+    3: "SBAS",
+    4: "BeiDou",
+    5: "Galileo",
+    6: "GLONASS",
+    7: "NavIC",
+}
+"""Information source for the future leap second event, from UBX-NAV-TIMELS"""